Test scanlines, fix lines_to_scanlines
lines_to_scanlines was creating a list of lists of scanlines. Each line
was a list of scanlines with all the letters combined. Then those lists
were put in a list. What needs to happen is that there's one list of
scanlines: all the scanlines from the lines of letters stacked on top of
each other.

Here's 4 letters that are 2x2 pixels laid out in a 2x2 grid (i.e. 2
letters each on two lines) with single integers as stand-ins for their
pixel RBGA values:

                      Letter-1   Letter-2
Line 1 - Scanline-S1,   1,  2,     5,  6
         Scanline-S2,   3,  4,     7,  8
Line 2 - Scanline-S3,   9, 10,    11, 12
         Scanline-S4,  13, 14,    15, 16

Laid out like this it easier to see that the scanlines should look like
this:

S1' [[ 1,  2,  5,  6]
S2'  [ 3,  4,  7,  8]
S3'  [ 9, 10, 11, 12]
S4'  [13, 14, 15, 16]]

The scanlines of the letters (e.g. 1,2,5,6 for the top scanline of
letters 1 and 2 on line 1) are grouped together in a single scanline.
The list of scanlines (S1 to S4) for lines 1 and 2 are then concatenated.