Extract rendered letter from "canvas"
erl_freetype passes a standard sized char array into the C code and
copies the rendered alpha map into that char array. The letter should
always be smaller than the array. When that array is passed back to
Erlang it becomes a binary. The binary, when considered as a cartesian
plane, is wider and taller than the letter. I need to "shrink" the
binary down to just big enough to hold the letter. erl_freetype
passes back the height and width of the rendered letter so I use that
to extract the letter from the binary. I break the binary into lines
using the original width of the standard sized char array and then from
those lines I use the rendered letter width to get the letter line. I do
this h times where h is the height of the letter. The origin of the
rendered letter is 0,0 wher y is going down, so I'm chopping off the
right-most and bottom-most portion of the binary that is unnecessary.

erl_to_png now writes a PNG file but it's garbage. I need to figure out
what's going wrong.